Fix race condition on parallel processes deleting .tmp directory - closes #71
