+++ lib/documentScreenshot.js
@@ -25,7 +25,9 @@
 var async = require('async'),
     fs = require('fs'),
     gm = require('gm'),
-    rimraf = require('rimraf');
+    rimraf = require('rimraf'),
+    generateUUID = require("./generateUUID.js"),
+    path = require('path');
 
 module.exports = function documentScreenshot(fileName) {
 
@@ -48,7 +50,7 @@ module.exports = function documentScreenshot(fileName) {
             execute: [],
             screenshot: []
         },
-        tmpDir = __dirname + '/../.tmp',
+        tmpDir = null,
         cropImages = [],
         currentXPos = 0,
         currentYPos = 0,
@@ -80,6 +82,9 @@ module.exports = function documentScreenshot(fileName) {
          * create tmp directory to cache viewport shots
          */
         function(cb) {
+            var uuid = generateUUID();
+            tmpDir = path.join(__dirname, '..', '.tmp-' + uuid);
+
             fs.exists(tmpDir, function(exists) {
                 return exists ? cb() : fs.mkdir(tmpDir, 0755, cb);
             });
