+++ lib/webdrivercss.js
@@ -11,6 +11,7 @@ var fs = require('fs-extra'),
     workflow = require('./workflow.js'),
     viewportScreenshot = require('./viewportScreenshot.js'),
     documentScreenshot = require('./documentScreenshot.js');
+    generateUUID = require("./generateUUID.js");
 
 /**
  * initialise plugin
@@ -21,19 +22,6 @@ var WebdriverCSS = function(webdriverInstance, options) {
     if(!webdriverInstance) {
         throw new Error('A WebdriverIO instance is needed to initialise WebdriverCSS');
     }
-
-    /**
-     * @return {string} A V4 UUID.
-     * @private
-     */
-    var _generateUUID = function (){
-        var d = new Date().getTime();
-        return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
-            var r = (d + Math.random()*16)%16 | 0;
-            d = Math.floor(d/16);
-            return (c=='x' ? r : (r&0x7|0x8)).toString(16);
-        });
-    };
 
     var that = this;
 
@@ -57,7 +45,7 @@ var WebdriverCSS = function(webdriverInstance, options) {
         apiKey: options.key,
         saveNewTests: true, // currently will always happen.
         saveFailedTests: this.updateBaseline,
-        batchId: _generateUUID()
+        batchId: generateUUID()
     };
     this.host = 'https://eyessdk.applitools.com';
     this.headers = {
